Quick Outlines
I pasted in some words that I had written in the other chef fundamentals repo that seemed appropriate here. I also filled in a few pieces from having run this class live in NYC. There are still a huge amount of edits necessary to make these the correct scripts for these scenes but this should give someone a good start.