Resources scene
First revisions performed on this sections initial script. I learned the best way to review this material is to actually run through it with the slides. This helped create this content.

Each body of text is now separated with a single dash to make it a little clearer what text belongs to what slides.