Updated Cookbooks (scene 04)
This is the initial draft of the script for the material.

The slides were updated to again use recipes that mirror the names of the cookbook and that is so that we have an exercise within the chef-client setcion.

I removed some the extra exercise slides that followed through the guided process.

I also added a slide to remind users to return to the home directory before creating an apache cookbook in the next exercise.

I added the applying of the apache as a step.

There may still be some pieces missing here which is taking a momement to view the hello world with curl.