chef-client - scene 05 updated
This section started out as a necessary evil. We had to get the attendees away from chef-apply and it seemed important that we did. Specifically before we got to templates or templating (which happens much later).

This at first felt awkward but what I decided is that this was a good moment to introduce a little bit of cookbook organization and include_recipe came to mind.

Teaching `include_recipe` here is more to get the attendees use to typing the `chef-client` command a number of times so that later we can say: Use chef-client to apply the COOKBOOK::RECIPE.