Scene 04 - Draft 2
Read aloud before intial recordings.

Added an additional slide on explaining the git staging area. This was mostly to give space to allow for the presenter to give the box analogy and provde links to some git reference materials. This is also one of the differences when coming from other source control tools like SVN.